main/requirements.txt
asgiref==3.8.1
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
crispy-bootstrap5==2024.10
cuser==0.0.4
Django==5.1.6
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-cuser==2017.3.16
django-environ==0.11.2
django-filter==25.1
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
drf-yasg==1.21.9
Flask==3.1.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
MarkupSafe==3.0.2
modeltranslation==0.25
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyserial==3.5
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
gunicorn
